Eliminación de Binarios por alerta GitHub
GitHub <support@github.com>

15:36 (hace 3 horas)

para mí
The page build completed successfully, but returned the following warning for the `master` branch:

It looks like you're using GitHub Pages to distribute binary files. We strongly suggest that you use releases to ship projects on GitHub. Releases are GitHub's way of packaging and providing software to your users. You can think of it as a replacement to using downloads to provide software. We found the following file(s) which may be a good candidate for releases: EllipseCalidadOTExcelAddIn/Instalador/setup.exe.